requirements.txt
asgiref==3.7.2
Django==4.2.2
django-debug-toolbar==4.1.0
django-render-partial==0.4
djangorestframework==3.14.0
pytz==2023.3
sqlparse==0.4.4
